+++ src/main/java/roomescape/payment/infrastructure/PaymentClient.java
@@ -1,7 +1,7 @@
 package roomescape.payment.infrastructure;
 
-import roomescape.payment.dto.request.PaymentCommand;
-import roomescape.payment.dto.response.PaymentResponse;
+import roomescape.payment.infrastructure.dto.reqeust.PaymentCommand;
+import roomescape.payment.infrastructure.dto.response.PaymentResponse;
 
 public interface PaymentClient {
     PaymentResponse authPayment(PaymentCommand paymentCommand);
