+++ src/main/java/roomescape/payment/infrastructure/TossApiClient.java
@@ -9,9 +9,9 @@ import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.HttpServerErrorException;
 import org.springframework.web.client.ResourceAccessException;
 import org.springframework.web.client.RestClient;
-import roomescape.payment.dto.request.PaymentCommand;
-import roomescape.payment.dto.request.TossPaymentRequest;
-import roomescape.payment.dto.response.PaymentResponse;
+import roomescape.payment.infrastructure.dto.reqeust.PaymentCommand;
+import roomescape.payment.infrastructure.dto.reqeust.TossPaymentRequest;
+import roomescape.payment.infrastructure.dto.response.PaymentResponse;
 import roomescape.payment.exception.TossPaymentClientException;
 import roomescape.payment.exception.TossPaymentServerException;
 import roomescape.payment.exception.TossServerErrorCode;
