+++ src/main/java/roomescape/payment/infrastructure/dto/response/PaymentResponse.java
@@ -1,4 +1,4 @@
-package roomescape.payment.dto.response;
+package roomescape.payment.infrastructure.dto.response;
 
 
 public record PaymentResponse(String paymentKey, String orderId, String type,
