+++ src/test/java/roomescape/payment/TossApiClientTest.java
@@ -17,8 +17,8 @@ import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.client.MockRestServiceServer;
 import org.springframework.web.client.RestClient;
-import roomescape.payment.dto.request.PaymentCommand;
-import roomescape.payment.dto.response.PaymentResponse;
+import roomescape.payment.infrastructure.dto.reqeust.PaymentCommand;
+import roomescape.payment.infrastructure.dto.response.PaymentResponse;
 import roomescape.payment.exception.TossPaymentClientException;
 import roomescape.payment.exception.TossPaymentServerException;
 import roomescape.payment.infrastructure.TossApiClient;
